+++ examplepatches/humzakh/HK-Morphe-Patches/patches/src/main/kotlin/app/template/patches/reddit/customclients/sync/syncforreddit/fix/otherdiscussions/Fingerprints.kt
@@ -6,6 +6,5 @@ import com.android.tools.smali.dexlib2.AccessFlags
 internal val otherDiscussionsBuildUrlFingerprint = Fingerprint(
     accessFlags = listOf(AccessFlags.PUBLIC, AccessFlags.STATIC),
     returnType = "Ljava/lang/String;",
-    strings = listOf("submit.json?url="),
-    custom = { _, classDef -> classDef.type == "Ly8/k0;" }
+    strings = listOf("submit.json?url=")
 )