+++ examplepatches/humzakh/HK-Morphe-Patches/patches/src/main/kotlin/app/template/patches/reddit/customclients/sync/syncforreddit/gestures/SyncPredictiveBackPatch.kt
@@ -3,11 +3,10 @@ package app.template.patches.reddit.customclients.sync.syncforreddit.gestures
 import app.morphe.patcher.patch.resourcePatch
 import app.template.patches.reddit.customclients.sync.syncforreddit.SyncForRedditCompatible
 
-@Suppress("unused")
 val syncPredictiveBackPatch = resourcePatch(
     name = "Predictive back gesture (Reddit Sync)",
     description = "Enables the predictive back gesture with fixes for Reddit Sync.",
-    default = true
+    default = false
 ) {
     compatibleWith(*SyncForRedditCompatible)