+++ examplepatches/kiraio-moe/Lain-Patches/patches/src/main/kotlin/kiraio/lain/photofancie/pro/Fingerprints.kt
@@ -1,0 +1,7 @@
+package kiraio.lain.photofancie.pro
+
+import app.morphe.patcher.Fingerprint
+
+object ProFingerprint : Fingerprint(
+    strings = listOf("NoProductPurchased")
+)