+++ examplepatches/kiraio-moe/Lain-Patches/CHANGELOG.md
@@ -1,3 +1,37 @@
+## [1.2.0](https://github.com/kiraio-moe/Lain-Patches/compare/v1.1.0...v1.2.0) (2026-06-28)
+
+### ✨ New Features
+
+* **AudioRelay:** Add premium patch ([e5cfc42](https://github.com/kiraio-moe/Lain-Patches/commit/e5cfc422468c15eccd23719bdb1845c5216849cd))
+* **Boorusama:** Add plus/premium patch ([e3daee2](https://github.com/kiraio-moe/Lain-Patches/commit/e3daee2df5fb02dc3048b1be281ae9292042de93))
+* **Hermit:** Add premium patch ([129d44e](https://github.com/kiraio-moe/Lain-Patches/commit/129d44eee62074b75386839173e2bab62fdbb3d3))
+* **Paletta:** Add premium patch ([f92f9c3](https://github.com/kiraio-moe/Lain-Patches/commit/f92f9c316de89bd0df4e8c3b6d90d7e229b48dce))
+* **PhotoFancie:** Add pro patch ([f80a4dd](https://github.com/kiraio-moe/Lain-Patches/commit/f80a4ddc407f28a862b1361b1b319da2c01d8070))
+* **Super Proxy:** Add pro patch ([6f4e784](https://github.com/kiraio-moe/Lain-Patches/commit/6f4e784d4c56dcfe33928352de9c0ba2f9d891ef))
+* **UDisc:** Add patches for UDisc ([d3629df](https://github.com/kiraio-moe/Lain-Patches/commit/d3629df87b7ae117d1763d820c01b2b96ea918c3))
+
+## [1.2.0-dev.3](https://github.com/kiraio-moe/Lain-Patches/compare/v1.2.0-dev.2...v1.2.0-dev.3) (2026-06-28)
+
+### ✨ New Features
+
+* **Boorusama:** Add plus/premium patch ([e3daee2](https://github.com/kiraio-moe/Lain-Patches/commit/e3daee2df5fb02dc3048b1be281ae9292042de93))
+* **PhotoFancie:** Add pro patch ([f80a4dd](https://github.com/kiraio-moe/Lain-Patches/commit/f80a4ddc407f28a862b1361b1b319da2c01d8070))
+* **Super Proxy:** Add pro patch ([6f4e784](https://github.com/kiraio-moe/Lain-Patches/commit/6f4e784d4c56dcfe33928352de9c0ba2f9d891ef))
+
+## [1.2.0-dev.2](https://github.com/kiraio-moe/Lain-Patches/compare/v1.2.0-dev.1...v1.2.0-dev.2) (2026-06-24)
+
+### ✨ New Features
+
+* **Hermit:** Add premium patch ([129d44e](https://github.com/kiraio-moe/Lain-Patches/commit/129d44eee62074b75386839173e2bab62fdbb3d3))
+* **UDisc:** Add patches for UDisc ([d3629df](https://github.com/kiraio-moe/Lain-Patches/commit/d3629df87b7ae117d1763d820c01b2b96ea918c3))
+
+## [1.2.0-dev.1](https://github.com/kiraio-moe/Lain-Patches/compare/v1.1.0...v1.2.0-dev.1) (2026-06-20)
+
+### ✨ New Features
+
+* **AudioRelay:** Add premium patch ([e5cfc42](https://github.com/kiraio-moe/Lain-Patches/commit/e5cfc422468c15eccd23719bdb1845c5216849cd))
+* **Paletta:** Add premium patch ([f92f9c3](https://github.com/kiraio-moe/Lain-Patches/commit/f92f9c316de89bd0df4e8c3b6d90d7e229b48dce))
+
 ## [1.1.0](https://github.com/kiraio-moe/Lain-Patches/compare/v1.0.0...v1.1.0) (2026-06-19)
 
 ### 🐛 Bug Fixes